Use initWithCoder: instead of awakeFromNib
When configuring an SMPageControl from inside a Storyboard or XIB file, the developer might want to use User Defined Runtime Attributes to configure the instance. awakeFromNib is called after those attributes are evaluated by the loading system, effectively overwriting the attributes defined in the Storyboard or XIB file. Therefore, initWithCoder: is the more appropriate location to perform the initialization.